fix(tutorials): restructured to fix circular imports re agentnetwork
