+++ agentMET4FOF_tutorials/tutorial_1_generator_agent.py
@@ -2,7 +2,8 @@ from time import sleep
 
 import numpy as np
 
-from agentMET4FOF.agents import AgentNetwork, MonitorAgent, SineGeneratorAgent
+from agentMET4FOF.network import AgentNetwork
+from agentMET4FOF.agents import MonitorAgent, SineGeneratorAgent
 
 
 def demonstrate_generator_agent_use() -> AgentNetwork:
