+++ agentMET4FOF_tutorials/tutorial_4_metrological_streams.py
@@ -1,4 +1,4 @@
-from agentMET4FOF.agents import AgentNetwork
+from agentMET4FOF.network import AgentNetwork
 from agentMET4FOF.metrological_agents import MetrologicalAgent, MetrologicalMonitorAgent
 from agentMET4FOF.metrological_streams import (
     MetrologicalDataStreamMET4FOF,
