+++ agentMET4FOF_tutorials/tutorial_5_coalition.py
@@ -1,4 +1,5 @@
-from agentMET4FOF.agents import AgentNetwork, MonitorAgent, SineGeneratorAgent
+from agentMET4FOF.agents import MonitorAgent, SineGeneratorAgent
+from agentMET4FOF.network import AgentNetwork
 
 
 def demonstrate_generator_agent_use():
