+++ agentMET4FOF_tutorials/tutorial_7_generic_metrological_agent.py
@@ -1,4 +1,4 @@
-from agentMET4FOF.agents import AgentNetwork
+from agentMET4FOF.network import AgentNetwork
 from agentMET4FOF.metrological_agents import (
     MetrologicalMonitorAgent,
     MetrologicalGeneratorAgent,
